+++ minimal_uhdless_demo/src/clock_sync.cpp
@@ -77,8 +77,8 @@ bool clock_sync::time_diff_recv(time_diff_resp & reply) {
 
         // Swap byte order from big to native
         // TODO: detect if we are using big or little endian at compile time, this currently assumes little endian
-        reply.tv_sec = __builtin_bswap64(reply.tv_sec);
-        reply.tv_tick = __builtin_bswap64(reply.tv_tick);
+        reply.tv_sec = static_cast<int64_t>(__builtin_bswap64(static_cast<uint64_t>(reply.tv_sec)));
+        reply.tv_tick = static_cast<int64_t>(__builtin_bswap64(static_cast<uint64_t>(reply.tv_tick)));
         return true;
 
     } else {
